# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API untuk platform OtoMeet.",
        "title": "OtoMeet API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/my-touring": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Melihat daftar agenda touring apa saja yang pernah diikuti oleh user login saat ini",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Riwayat Touring Saya",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Registration"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
...
        "/api/register-touring": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mendaftarkan diri dan kendaraan ke dalam salah satu jadwal touring",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Gabung Agenda Touring",
                "parameters": [
                    {
                        "description": "Data Kendaraan",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RegisterTouringRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
...
        "/api/touring/{id}/peserta": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin dapat melihat semua peserta yang mendaftar pada suatu agenda touring",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Daftar Peserta Touring",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Touring ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Registration"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger401Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Swagger403Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerBasicResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.RegisterTouringRequest": {
            "type": "object",
            "properties": {
                "bike_model": {
                    "type": "string",
                    "example": "Yamaha NMAX 155"
                },
                "license_plate": {
                    "type": "string",
                    "example": "D 1234 ABC"
                },
                "touring_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.Registration": {
            "type": "object",
            "properties": {
                "bike_model": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_plate": {
                    "type": "string"
                },
                "registered_at": {
                    "type": "string"
                },
                "touring": {
                    "$ref": "#/definitions/model.Touring"
                },
                "touring_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Touring": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deskripsi": {
                    "type": "string"
                },
                "harga_tiket": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "kuota": {
                    "type": "integer"
                },
                "lokasi_akhir": {
                    "type": "string"
                },
                "lokasi_awal": {
                    "type": "string"
                },
                "nama_touring": {
                    "type": "string"
                },
                "tanggal": {
                    "type": "string"
                },
                "tujuan": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "waktu": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "Changed: removed NOT NULL to handle existing data",
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "profile_photo": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "vehicles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Vehicle"
                    }
                }
            }
        },
        "model.Vehicle": {
            "type": "object",
            "properties": {
                "bike_model": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_plate": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "utils.Swagger401Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "token tidak valid atau sudah expired"
                }
            }
        },
        "utils.Swagger403Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "user tidak memiliki akses untuk fitur ini"
                }
            }
        },
        "utils.SwaggerBasicResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "string"
                },
                "error": {
                    "type": "string",
                    "example": "detail error"
                },
                "message": {
                    "type": "string",
                    "example": "detail pesan"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}